More operations, more complete buoyancy demo
Added more cpShape getters and setters.
Added cpBBNew.
Added list for shapes used in buoyancy demo, as well as started putting
in opeations to define how those shapes will behave.
NOT_GRABABLE_FLAG breaking in cpShapeSetLayers call, need to coerce to a
uint32 type so that it does not attempt to pass a negative value.
Definition for the mask should be equivalent to the Chipmunk mask,
according to how I've seen it defined online. Sans the fact that it's a
plain int right now.