Buoyancy Revision 1
Fully complete translation of C to Racket. However, it doesn't actually
run and crashes with an incorrect number of arguments error. Need to do
some debugging to resolve those and make sure the behavior works.
Also need to get program to draw all shapes, right now will only draw
one of the rectangles (when it works).