No Set/GetIterations in libchipmunk. io _ptrs
Removed bindings to functions that don't actually exist in libchipmunk,
making it so that you could not use the bindings.
Changed pointers to use io, since we need to use them after the call to
a function. (I forget what io does EXACTLY, but I remember the Racket
IRC saying that they were needed for this use case)