Attempted fix for double pointer crashes Racket
I still don't really know how to do this, and have been experimenting
with it for a few hours as well as re-attempting a suggestion I got a
while back. All attempts so far have just caused Racket to either reject
a pointer type, or for the current code, completely crash Racket.
Crash happens on line 19 with cpPolyShapeNumVerts, where I'm currently
assuming chipmunk is trying to access something inside of the cpShape,
which probably has not been actually initialized.
A possible reason: the cpArbiter that is getting passed in to the water
presolver is not fully initialized in some way, causing the shapes that
cpArbiterGetShapes  uses to not actually set a real object into the
pointer. Or maybe I'm just not passning the correct pointer type to
cpArbiterGetShapes or cpPolyShapeGetNumVerts? Wrong pointers seems more
likely.